backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["gatlace <zii.donato@outlook.com>"]

[tool.poetry.scripts]


[tool.poetry.dependencies]
python = "^3.10"
Django = "^4.0.6"
psycopg2 = "^2.9.3"
black = "^22.6.0"
xcffib = "^0.11.1"
djangorestframework = "^3.13.1"
Faker = "^13.15.1"
requests = "^2.28.1"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
